+++ spo/spo/report/anstehende_beratungen/anstehende_beratungen.json
@@ -8,7 +8,7 @@
  "idx": 0,
  "is_standard": "Yes",
  "json": "{}",
- "modified": "2021-12-15 14:49:36.832900",
+ "modified": "2022-08-02 14:49:36.832900",
  "modified_by": "Administrator",
  "module": "SPO",
  "name": "Anstehende Beratungen",
@@ -22,7 +22,10 @@
    "role": "System Manager"
   },
   {
-   "role": "Sales Manager"
+   "role": "SPO Poweruser"
+  },
+  {
+   "role": "Beratung"
   }
  ]
 }
