--- spo/spo/report/gebuchte_slots/gebuchte_slots.js
@@ -1,9 +1,0 @@
-// Copyright (c) 2022, libracore and contributors
-// For license information, please see license.txt
-/* eslint-disable */
-
-frappe.query_reports["Gebuchte Slots"] = {
-	"filters": [
-
-	]
-};
