--- spo/spo/report/gebuchte_slots/gebuchte_slots.json
@@ -1,30 +1,0 @@
-{
- "add_total_row": 0,
- "creation": "2022-07-07 15:37:42.617284",
- "disable_prepared_report": 0,
- "disabled": 0,
- "docstatus": 0,
- "doctype": "Report",
- "idx": 0,
- "is_standard": "Yes",
- "modified": "2022-07-07 15:37:42.617284",
- "modified_by": "Administrator",
- "module": "SPO",
- "name": "Gebuchte Slots",
- "owner": "Administrator",
- "prepared_report": 0,
- "ref_doctype": "Beratungsslot",
- "report_name": "Gebuchte Slots",
- "report_type": "Script Report",
- "roles": [
-  {
-   "role": "System Manager"
-  },
-  {
-   "role": "SPO Poweruser"
-  },
-  {
-   "role": "Beratung"
-  }
- ]
-}
