+++ spo/utils/cleanup.py
@@ -28,26 +28,47 @@ def cleanup_slots():
     date = add_days(nowdate(), -10)
     slots = frappe.db.sql("""SELECT `name` FROM `tabBeratungsslot` WHERE `status` = 'reserviert' AND `tabBeratungsslot`.`start` <= '{date}' """.format(date=date), as_dict=True)
     for slot in slots:
-        slot = frappe.get_doc("Beratungsslot", slot.name)
+        slot_doc = frappe.get_doc("Beratungsslot", slot['name'])
         
-        # cancel invoice
-        invoices = frappe.db.sql("""SELECT `name` FROM `tabSales Invoice` WHERE `docstatus` = 1 AND `beratungsslot` = '{slot}'""".format(slot=slot.name), as_dict=True)
-        for sinv in invoices:
-            sinv = frappe.get_doc("Sales Invoice", sinv.name)
-            sinv.cancel()
-        
-        # deactivate customer
-        customer = frappe.get_doc("Customer", slot.customer)
-        customer.disabled = 1
-        customer.save()
+        # enable customer (otherwise invoice cannot be cancelled)
+        if frappe.db.exists("Customer", slot_doc.customer):
+            customer = frappe.get_doc("Customer", slot_doc.customer)
+            customer.disabled = 0
+            try:
+                customer.save()
+            except Exception as err:
+                frappe.log_error(err, "cleanup_slots: enable customer {0}".format(slot_doc.customer))
+            
+            # cancel invoice
+            invoices = frappe.db.sql("""SELECT `name` FROM `tabSales Invoice` WHERE `docstatus` = 1 AND `beratungsslot` = '{slot}'""".format(slot=slot_doc.name), as_dict=True)
+            for sinv in invoices:
+                sinv_doc = frappe.get_doc("Sales Invoice", sinv['name'])
+                try:
+                    sinv_doc.cancel()
+                except Exception as err:
+                    frappe.log_error(err, "cleanup_slots: cancel invoice {0}".format(sinv_doc.name))
+            
+            # deactivate customer
+            customer.disabled = 1
+            try:
+                customer.save()
+            except Exception as err:
+                frappe.log_error(err, "cleanup_slots: disable customer {0}".format(slot_doc.customer))
         
         # set slot status
-        slot.status = 'storniert'
-        slot.save()
+        slot_doc.status = 'storniert'
+        try:
+            slot_doc.save()
+        except Exception as err:
+            frappe.log_error(err, "cleanup_slots: mark slot {0}".format(slot_doc.name))
     
     # delete old unused slots
     date = add_days(nowdate(), -7)
     slots = frappe.db.sql("""SELECT `name` FROM `tabBeratungsslot` WHERE `status` = 'frei' AND `tabBeratungsslot`.`start` <= '{date}' """.format(date=date), as_dict=True)
     for slot in slots:
-        slot = frappe.get_doc("Beratungsslot", slot.name)
-        slot.delete()
+        slot_doc = frappe.get_doc("Beratungsslot", slot['name'])
+        try:
+            slot_doc.delete()
+        except Exception as err:
+            frappe.log_error(err, "cleanup_slots: remove slot {0}".format(slot.customer))
+    return
